Starting point for coding the nunjucks loader.
