+++ lib/dev-server.js
@@ -110,9 +110,11 @@ function runNodemon (port) {
     watch: [
       path.join(projectDir, '.env'),
       path.join(appDir, '**', '*.js'),
-      path.join(appDir, '**', '*.json')
+      path.join(appDir, '**', '*.json'),
+      path.join(projectDir, 'node_modules', 'govuk-prototype-kit', '**', '*.js')
     ],
     script: path.join(packageDir, 'listen-on-port.js'),
+    ignoreRoot: [".git"],
     ignore: [
       'app/assets/*'
     ]
