+++ server.js
@@ -8,7 +8,7 @@ const bodyParser = require('body-parser')
 const cookieParser = require('cookie-parser')
 const dotenv = require('dotenv')
 const express = require('express')
-const { expressNunjucks, getNunjucksAppEnv } = require('./lib/nunjucks/nunjucksLoader')
+const { expressNunjucks, getNunjucksAppEnv } = require('./lib/nunjucks/nunjucksConfiguration')
 
 // We want users to be able to keep api keys, config variables and other
 // envvars in a `.env` file, run dotenv before other code to make sure those
@@ -72,7 +72,7 @@ app.use(sessionUtils.getSessionMiddleware())
 // Set up App
 const appViews = [
   path.join(projectDir, '/app/views/')
-].concat(plugins.getAppViews([backupNunjucksDir, finalBackupNunjucksDir]))
+].concat(plugins.getAppViews())
 
 const nunjucksConfig = {
   autoescape: true,
